优化技能描述，适配 actionType 132、133

+++ app/src/main/java/cn/wthee/pcrtool/data/db/view/skilltype/StrikeBack.kt
@@ -29,7 +29,7 @@ fun SkillActionDetail.strikeBack(): String {
      */
     val action = if (actionValue3.toInt() != 0) {
         // 动作
-        getString(R.string.skill_action_type_desc_33_action, actionDetail3.toInt() % 100)
+        getString(R.string.skill_action_type_desc_33_action, actionDetail3 % 100)
     } else {
         // 数值
         getString(R.string.skill_action_type_desc_33_value, value)
